chore(core): Move out deepEqual, getShaderInfo
