+++ modules/core/src/utils/uid.ts
@@ -14,15 +14,3 @@ export function uid(id: string = 'id'): string {
   const count = uidCounters[id]++;
   return `${id}-${count}`;
 }
-
-/** Returns true if given object is empty, false otherwise. */
-export function isObjectEmpty(obj: object): boolean {
-  let isEmpty = true;
-  // @ts-ignore key is unused
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  for (const key in obj) {
-    isEmpty = false;
-    break;
-  }
-  return isEmpty;
-}
