+++ modules/core/test/index.ts
@@ -1,23 +1,20 @@
 // general utils
 import './lib/utils/array-utils-flat.spec';
-import './lib/utils/utils.spec';
-import './lib/utils/deep-equal.spec';
-import './lib/utils/uniform.spec';
+import './lib/utils/uid.spec';
 
 // type utils
-import './adapter/attribute-utils/get-attribute-from-layout.spec';
-
 import './adapter/type-utils/decode-attribute-type.spec';
-
 import './adapter/type-utils/decode-vertex-format.spec';
 import './adapter/type-utils/decode-texture-format.spec';
 import './adapter/type-utils/vertex-format-from-attribute.spec';
 
 // adapter
+import './adapter/attribute-utils/get-attribute-from-layout.spec';
 import './adapter/canvas-context.spec';
 
 // uniforms
 import './lib/uniforms/uniform-buffer-layout.spec';
+import './lib/uniforms/uniform.spec';
 
 // compiler logs
 import './lib/compiler-log/format-compiler-log.spec';
