+++ modules/engine/test/index.ts
@@ -1,3 +1,7 @@
+// utils
+import './lib/utils/deep-equal.spec';
+
+// model etc
 import './lib/model.spec';
 import './lib/animation-loop.spec';
 import './lib/pipeline-factory.spec';
